fix(utxo/postgres): align reorg/atomicity/freeze/ownership with aerospike gold standard
Address PR #684 review findings, using the aerospike store as the authoritative
reference for expected behaviour.

Critical / high (consensus, data-loss):
- unsetMinedMulti (reorg) now clears delete_at_height + locked when a tx falls off
  the longest chain, and rewinds the DAH sweep watermark to the reorged height.
  Previously a reorged-out, fully-spent tx kept its stale prune stamp and was
  deleted by the pruner, destroying still-live UTXOs.
- Unspend is now atomic (delete + DAH-clear + lock in one transaction) and
  ownership-checked via spending_data (nil token rejected), matching the
  aerospike/sql stores. A non-owning caller no longer wipes a live spend.
- FreezeUTXOs no longer sets the transaction-level frozen flag; freezing one
  output of a multi-output tx no longer blocks spending its siblings.
- SetMinedMulti enforces the all-hashes postcondition (TxNotFound on a missing
  hash) and checks rows.Err(); mined_at_height is bound as a parameter (block
  height) instead of an interpolated stale chain tip.
- Pruner cascade DELETE re-checks delete_at_height at execution time, closing the
  race where a concurrent reorg Unspend revives a tx between the tombstone scan
  and the delete.

Lower severity:
- get.go batchDecorateChunk checks rows.Err() (no more spurious TxNotFound on a
  truncated result set).
- DAH sweep + backstop full-spend test counts only spends of spendable outputs.
- DAH watermark advance is forward-only under concurrent sweepers.
- Background DAH cursor runs under a store-owned context (Stop/Close cancel it).
- conflicting.go checks RowsAffected on the conflicting UPDATE.
- int64-widened DAH height arithmetic (delete.go, conflicting.go, preservation.go).

ProcessExpiredPreservations intentionally keeps aerospike's unconditional DAH
stamping on expiry (the gold standard's safety gate is the pruner, not stamp-time
eligibility).

Adds regression tests: per-output freeze isolation, Unspend ownership, reorg
DAH-clear. Full postgres suite passes with -race.

+++ stores/utxo/postgres/delete.go
@@ -85,7 +85,9 @@ func (s *Store) setDAH(ctx context.Context, hash *chainhash.Hash) error {
 	}
 
 	hasBlockIDs := len(blockIDs) > 0
-	newDAH := int64(s.blockHeight.Load() + 1 + retention)
+	// Widen to int64 BEFORE adding so the sum cannot wrap in uint32 arithmetic
+	// (blockHeight and retention are both uint32). Matches dah_sweep.go.
+	newDAH := int64(s.blockHeight.Load()) + 1 + int64(retention)
 
 	if allSpent && hasBlockIDs && onLongestChain {
 		// Set delete_at_height, but only bump forward (never decrease).
